Fixes Issue When Merging Templates
If a default template has a nested key and the top template has a
templated base key, then, when merging, the top template key will
replace the entire map of the base template. This is not desired as the
base template may have certain keys that don't collide and that we want
to preserve.