Fixes Logic Errors with Merge
We need to first populate all templates then perform a merge operation.
This is the case because we need nested keys from both templates to be
considered when identifying a collision and replacement. Otherwise,
nested keys are entirely removed if a parentt key is templated, which
isn't what we want.