Add X-Cadre-Secret Authorization Flow (#17)
* Add X-Cadre-Secret Authorization Flow

Cadre can now be started with a secret token parameter. This allows for
the server to check for that token on every request (except `GET
/ping`), which provides a minimum layer of security for the server and
clients.

One can pass the secret token as a CLI flag (`--secret`) or via env var
(`CADRE_SECRET`).

* Adds token to frontend

Templates the frontend with handlebars which are then populated using
`tera` and the auth secret. This assumes that the frontend is only
accessible internally.

* Revert "Adds token to frontend"

This reverts commit e68b5db6b8132dccffd52a56b62be9bf7e1bc03d.

* Removes #allow

* Makes secret a requirement

Cadre can no longer start without a secret.

* Makes secret mandatory

Instead of an optional propery, Cadre now requires a secret in order to
work. It won't accept any requests (except to render the UI and /ping)
without a secret.

* Adds input for secret in frontend

The frontend now takes an input for a secret. That input can also be
passed via parameter in URL.

* Fixes formatting